// __generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type Page_Grooming = AcfFieldGroup & {
  __typename?: 'Page_Grooming';
  description?: Maybe<Scalars['String']['output']>;
  descriptionAtFooter?: Maybe<Scalars['String']['output']>;
  faqs?: Maybe<Page_Grooming_Faqs>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Mỗi tấm hình phải xuống dòng */
  images?: Maybe<Scalars['String']['output']>;
  procedure?: Maybe<Page_Grooming_Procedure>;
  whyWeShouldUseGrooming?: Maybe<Scalars['String']['output']>;
};
// ...
export type Page_Grooming_Procedure = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  normalDog?: Maybe<Page_Grooming_Procedure_NormalDog>;
  unnormalDog?: Maybe<Page_Grooming_Procedure_UnnormalDog>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  step1?: Maybe<Page_Grooming_Procedure_NormalDog_Step1>;
  step2?: Maybe<Page_Grooming_Procedure_NormalDog_Step2>;
  step3?: Maybe<Page_Grooming_Procedure_NormalDog_Step3>;
  step4?: Maybe<Page_Grooming_Procedure_NormalDog_Step4>;
  step5?: Maybe<Page_Grooming_Procedure_NormalDog_Step5>;
  step6?: Maybe<Page_Grooming_Procedure_NormalDog_Step6>;
  step7?: Maybe<Page_Grooming_Procedure_NormalDog_Step7>;
  step8?: Maybe<Page_Grooming_Procedure_NormalDog_Step8>;
  step9?: Maybe<Page_Grooming_Procedure_NormalDog_Step9>;
  step10?: Maybe<Page_Grooming_Procedure_NormalDog_Step10>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step1 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step1';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step2 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step2';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step3 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step3';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step4 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step4';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step5 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step5';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step6 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step6';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step7 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step7';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step8 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step8';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step9 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step9';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_NormalDog_Step10 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_NormalDog_Step10';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Grooming_Procedure_UnnormalDog = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_UnnormalDog';
  afterNote?: Maybe<Scalars['String']['output']>;
  beforeNote?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  step1?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step1>;
  step2?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step2>;
  step3?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step3>;
  step4?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step4>;
  step5?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step5>;
  step6?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step6>;
  step7?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step7>;
  step8?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step8>;
  step9?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step9>;
  step10?: Maybe<Page_Grooming_Procedure_UnnormalDog_Step10>;
};

/** Field Group */
export type Page_Grooming_Procedure_UnnormalDog_Step1 = AcfFieldGroup & {
  __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step1';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type RootQueryToPageConnection = Connection & PageConnection & {
  __typename?: 'RootQueryToPageConnection';
  /** Edges for the RootQueryToPageConnection connection */
  edges: Array<RootQueryToPageConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Page>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToPageConnectionPageInfo;
};
// ...
export type Grooming_FragmentFragment = { __typename?: 'Page', grooming?: { __typename?: 'Page_Grooming', description?: string | null, whyWeShouldUseGrooming?: string | null, images?: string | null, descriptionAtFooter?: string | null, procedure?: { __typename?: 'Page_Grooming_Procedure', description?: string | null, normalDog?: { __typename?: 'Page_Grooming_Procedure_NormalDog', step1?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step1', title?: string | null, description?: string | null } | null, step2?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step2', title?: string | null, description?: string | null } | null, step3?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step3', title?: string | null, description?: string | null } | null, step4?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step4', title?: string | null, description?: string | null } | null, step5?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step5', title?: string | null, description?: string | null } | null, step6?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step6', title?: string | null, description?: string | null } | null, step7?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step7', title?: string | null, description?: string | null } | null, step8?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step8', title?: string | null, description?: string | null } | null, step9?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step9', title?: string | null, description?: string | null } | null, step10?: { __typename?: 'Page_Grooming_Procedure_NormalDog_Step10', title?: string | null, description?: string | null } | null } | null, unnormalDog?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog', beforeNote?: string | null, afterNote?: string | null, step1?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step1', title?: string | null, description?: string | null } | null, step2?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step2', title?: string | null, description?: string | null } | null, step3?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step3', title?: string | null, description?: string | null } | null, step4?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step4', title?: string | null, description?: string | null } | null, step5?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step5', title?: string | null, description?: string | null } | null, step6?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step6', title?: string | null, description?: string | null } | null, step7?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step7', title?: string | null, description?: string | null } | null, step8?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step8', title?: string | null, description?: string | null } | null, step9?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step9', title?: string | null, description?: string | null } | null, step10?: { __typename?: 'Page_Grooming_Procedure_UnnormalDog_Step10', title?: string | null, description?: string | null } | null } | null } | null, faqs?: { __typename?: 'Page_Grooming_Faqs', faq1?: { __typename?: 'Page_Grooming_Faqs_Faq1', question?: string | null, answer?: string | null } | null, faq2?: { __typename?: 'Page_Grooming_Faqs_Faq2', question?: string | null, answer?: string | null } | null, faq3?: { __typename?: 'Page_Grooming_Faqs_Faq3', question?: string | null, answer?: string | null } | null, faq4?: { __typename?: 'Page_Grooming_Faqs_Faq4', question?: string | null, answer?: string | null } | null, faq5?: { __typename?: 'Page_Grooming_Faqs_Faq5', question?: string | null, answer?: string | null } | null, faq6?: { __typename?: 'Page_Grooming_Faqs_Faq6', question?: string | null, answer?: string | null } | null, faq7?: { __typename?: 'Page_Grooming_Faqs_Faq7', question?: string | null, answer?: string | null } | null, faq8?: { __typename?: 'Page_Grooming_Faqs_Faq8', question?: string | null, answer?: string | null } | null } | null } | null } & { ' $fragmentName'?: 'Grooming_FragmentFragment' };
// ...
export type Get_GroomingQueryVariables = Exact<{
  name: Scalars['String']['input'];
}>;


export type Get_GroomingQuery = { __typename?: 'RootQuery', pages?: { __typename?: 'RootQueryToPageConnection', nodes: Array<(
      { __typename?: 'Page' }
      & { ' $fragmentRefs'?: { 'Fragment_Page_TypeFragment': Fragment_Page_TypeFragment;'Grooming_FragmentFragment': Grooming_FragmentFragment } }
    )> } | null };
// ...
export type Fragment_Page_TypeFragment = { __typename?: 'Page', featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: (
      { __typename?: 'MediaItem' }
      & { ' $fragmentRefs'?: { 'Fragment_Media_ItemFragment': Fragment_Media_ItemFragment } }
    ) } | null, seo?: { __typename?: 'PostTypeSEO', title?: string | null, metaDesc?: string | null } | null } & { ' $fragmentName'?: 'Fragment_Page_TypeFragment' };